split moddag into two files for readability
